+++ lib/browser/base.rb
@@ -197,6 +197,18 @@ module Browser
       ua =~ /SamsungBrowser/ && detect_version?(full_version, expected_version)
     end
 
+    # Detect if browser is Huawei.
+    def huawei_browser?(expected_version = nil)
+      HuaweiBrowser.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+
+    # Detect if browser is Xiaomi Miui.
+    def miui_browser?(expected_version = nil)
+      MiuiBrowser.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+
     def maxthon?(expected_version = nil)
       Maxthon.new(ua).match? && detect_version?(full_version, expected_version)
     end