+++ lib/browser/browser.rb
@@ -32,6 +32,8 @@ require_relative "sputnik"
 require_relative "snapchat"
 require_relative "duck_duck_go"
 require_relative "samsung_browser"
+require_relative "huawei_browser"
+require_relative "miui_browser"
 require_relative "maxthon"
 require_relative "mail_master"
 require_relative "google_search_app"
@@ -78,6 +80,8 @@ module Browser
       Sputnik,              # must be placed before Chrome and Safari
       DuckDuckGo,           # must be placed before Chrome and Safari
       SamsungBrowser,       # must be placed before Chrome and Safari
+      HuaweiBrowser,        # must be placed before Chrome and Safari
+      MiuiBrowser,          # must be placed before Chrome and Safari
       Maxthon,              # must be placed before Chrome and Safari
       MailMaster,           # must be placed before Chrome and Safari
       GoogleSearchApp,      # must be placed before Chrome and Safari