+++ lib/browser/safari.rb
@@ -23,6 +23,8 @@ module Browser
         !edge? &&
         !chrome? &&
         !samsung_browser? &&
+        !huawei_browser? &&
+        !miui_browser? &&
         !duck_duck_go? &&
         !yandex? &&
         !sputnik? &&